+++ tests/screenshot/apps/components/SwitchItem.js
@@ -21,6 +21,16 @@ const SwitchItemTests = [
 		<SwitchItem disabled>Hello SwitchItem</SwitchItem>,
 		// with icon
 		<SwitchItem icon="music">Hello SwitchItem</SwitchItem>
+	]),
+
+	// Focus
+	...withConfig({focus: true}, [
+		<SwitchItem>Hello focused SwitchItem</SwitchItem>,
+		<SwitchItem selected>Hello focused SwitchItem</SwitchItem>,
+		<SwitchItem disabled>Hello focused SwitchItem</SwitchItem>,
+
+		// with icon
+		<SwitchItem icon="music">Hello focused SwitchItem</SwitchItem>
 	])
 ];
 export default SwitchItemTests;