+++ tests/ui/specs/Drawer/Drawer-specs.js
@@ -23,7 +23,6 @@ describe('Drawer', function () {
 
 		it('should have correct heading', async function () {
 			await drawerCommon.buttonDrawer1.click();
-
 			await Page.waitForOpen(drawer);
 
 			await expectOpen(drawerCommon);
@@ -32,7 +31,6 @@ describe('Drawer', function () {
 
 		it('should not have a line under the heading text', async function () {
 			await drawerCommon.buttonDrawer1.click();
-
 			await Page.waitForOpen(drawer);
 
 			await expectOpen(drawerCommon);
@@ -48,10 +46,9 @@ describe('Drawer', function () {
 				expect(await drawer.isOpen).to.be.true();
 			});
 
-			it('should close drawer with cancel button on 5-way right in drawer container', async function () {
+			it('should close drawer with cancel button on 5-way right in drawer container', async function () {
 				await Page.spotlightSelect();
 				await Page.waitForOpen(drawer);
-				await browser.pause(400);
 
 				await expectOpen(drawerCommon);
 
@@ -59,7 +56,6 @@ describe('Drawer', function () {
 				await Page.spotlightSelect();
 
 				await Page.waitForClose(drawer);
-				await browser.pause(400);
 
 				await expectClosed(drawerCommon);
 			});
@@ -78,11 +74,9 @@ describe('Drawer', function () {
 				await Page.waitForOpen(drawer);
 
 				await expectOpen(drawerCommon);
-				await browser.pause(400);
 
 				await drawer.buttonCancel.click();
 				await Page.waitForClose(drawer);
-				await browser.pause(400);
 
 				await expectClosed(drawerCommon);
 			});
@@ -111,7 +105,7 @@ describe('Drawer', function () {
 				expect(await drawer.isOpen).to.be.true();
 			});
 
-			it('should close drawer with cancel button on 5-way right in drawer container', async function () {
+			it('should close drawer with cancel button on 5-way right in drawer container', async function () {
 				await Page.spotlightRight();
 				await Page.spotlightSelect();
 
@@ -121,7 +115,6 @@ describe('Drawer', function () {
 				await Page.spotlightSelect();
 				await Page.waitForClose(drawer);
 
-				await browser.pause(400);
 				await expectClosed(drawerCommon);
 			});
 		});
@@ -173,12 +166,11 @@ describe('Drawer', function () {
 				expect(await drawer.isOpen).to.be.true();
 			});
 
-			it('should close drawer with cancel button on 5-way right in drawer container', async function () {
+			it('should close drawer with cancel button on 5-way right in drawer container', async function () {
 				await Page.spotlightRight();
 				await Page.spotlightRight();
 				await Page.spotlightSelect();
 				await Page.waitForOpen(drawer);
-				await browser.pause(400);
 
 				await expectOpen(drawerCommon);
 
@@ -186,7 +178,6 @@ describe('Drawer', function () {
 				await Page.spotlightSelect();
 
 				await Page.waitForClose(drawer);
-				await browser.pause(400);
 
 				await expectClosed(drawerCommon);
 			});
@@ -209,7 +200,6 @@ describe('Drawer', function () {
 
 				await drawer.buttonCancel.click();
 				await Page.waitForClose(drawer);
-				await browser.pause(400);
 
 				await expectClosed(drawerCommon);
 			});
@@ -241,20 +231,18 @@ describe('Drawer', function () {
 				expect(await drawer.isOpen).to.be.true();
 			});
 
-			it('should close drawer with cancel button on 5-way right in drawer container', async function () {
+			it('should close drawer with cancel button on 5-way right in drawer container', async function () {
 				await Page.spotlightRight();
 				await Page.spotlightRight();
 				await Page.spotlightRight();
 				await Page.spotlightSelect();
 				await Page.waitForOpen(drawer);
-				await browser.pause(400);
 
 				await expectNoneScrimOpen(drawerCommon);
 
 				await Page.spotlightRight();
 				await Page.spotlightSelect();
 				await Page.waitForClose(drawer);
-				await browser.pause(400);
 
 				await expectClosed(drawerCommon);
 			});
@@ -271,13 +259,11 @@ describe('Drawer', function () {
 			it('should close the popup on click in drawer container', async function () {
 				drawerCommon.buttonDrawer4.click();
 				await Page.waitForOpen(drawer);
-				await browser.pause(400);
 
 				await expectNoneScrimOpen(drawerCommon);
 
 				await drawer.buttonOK.click();
 				await Page.waitForClose(drawer);
-				await browser.pause(400);
 
 				await expectClosed(drawerCommon);
 			});
@@ -290,7 +276,6 @@ describe('Drawer', function () {
 
 				await drawer.buttonCancel.click();
 				await Page.waitForClose(drawer);
-				await browser.pause(400);
 
 				await expectClosed(drawerCommon);
 			});